+++ jcodemodel/src/main/java/com/helger/jcodemodel/writer/options/wrap/Method.java
@@ -1,21 +1,21 @@
 package com.helger.jcodemodel.writer.options.wrap;
 
-import com.helger.jcodemodel.writer.options.Wrap.EWrapListStrategy;
-import com.helger.jcodemodel.writer.options.Wrap.EWrapWordStrategy;
-import com.helger.jcodemodel.writer.options.Wrap.WrapListMode;
-import com.helger.jcodemodel.writer.options.Wrap.WrapWordMode;
+import com.helger.jcodemodel.writer.options.Wrap.WrapList;
+import com.helger.jcodemodel.writer.options.Wrap.WrapList.EWrapListStrategy;
+import com.helger.jcodemodel.writer.options.Wrap.WrapWord;
+import com.helger.jcodemodel.writer.options.Wrap.WrapWord.EWrapWordStrategy;
 
 public class Method {
 
   /// declaration of the method's params
-  public final WrapListMode params =
-      new WrapListMode()
+  public final WrapList params =
+      new WrapList()
           .condition(EWrapListStrategy.PAST3)
           .indent(1);
 
   /// wrapping of opening bracket ( '{' )
-  public final WrapWordMode bracket =
-      new WrapWordMode()
+  public final WrapWord bracket =
+      new WrapWord()
           .condition(EWrapWordStrategy.NEVER)
           .indent(0);
